[CL-2438] Updates from code review
