+++ front/app/components/FormBuilder/components/FormBuilderSettings/ConfigSelectWithLocaleSwitcher/messages.ts
@@ -2,15 +2,15 @@ import { defineMessages } from 'react-intl';
 
 export default defineMessages({
   removeAnswer: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.removeAnswer',
+    id: 'app.components.formBuilder.removeAnswer',
     defaultMessage: 'Remove answer',
   },
   addAnswer: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.addAnswer',
+    id: 'app.components.formBuilder.addAnswer',
     defaultMessage: 'Add answer',
   },
   fieldLabel: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.fieldLabel',
+    id: 'app.components.formBuilder.fieldLabel',
     defaultMessage: 'Answer choices',
   },
 });
