+++ front/app/components/FormBuilder/components/FormBuilderSettings/LinearScaleSettings/messages.ts
@@ -2,24 +2,24 @@ import { defineMessages } from 'react-intl';
 
 export default defineMessages({
   labels: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.labels',
+    id: 'app.components.formBuilder.labels',
     defaultMessage: 'Labels',
   },
   labelsTooltipContent: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.labelsTooltipContent',
+    id: 'app.components.formBuilder.labelsTooltipContent',
     defaultMessage:
       'Choose optional labels for your minimum and maximum values.',
   },
   range: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.range',
+    id: 'app.components.formBuilder.range',
     defaultMessage: 'Range',
   },
   selectRangeTooltip: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.selectRangeTooltip',
+    id: 'app.components.formBuilder.selectRangeTooltip',
     defaultMessage: 'Choose the maximum value for your scale.',
   },
   toLabel: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.toLabel',
+    id: 'app.components.formBuilder.toLabel',
     defaultMessage: 'to',
   },
 });
