+++ front/app/components/FormBuilder/components/FormBuilderSettings/PageSettings/messages.ts
@@ -2,11 +2,11 @@ import { defineMessages } from 'react-intl';
 
 export default defineMessages({
   titleLabel: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.page.title',
+    id: 'app.components.formBuilder.page.title',
     defaultMessage: 'Title (optional)',
   },
   descriptionLabel: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.page.description',
+    id: 'app.components.formBuilder.page.description',
     defaultMessage: 'Description (optional)',
   },
 });
