+++ front/app/components/FormBuilder/components/FormFields/index.tsx
@@ -57,7 +57,7 @@ interface FormFieldsProps {
   ) => void;
   isEditingDisabled: boolean;
   selectedFieldId?: string;
-  builderConfig: FormBuilderConfig | undefined;
+  builderConfig: FormBuilderConfig;
 }
 
 const FormFields = ({
