+++ front/app/components/FormBuilder/messages.ts
@@ -2,24 +2,24 @@ import { defineMessages } from 'react-intl';
 
 export default defineMessages({
   emptyTitleError: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.emptyTitleError',
+    id: 'app.components.formBuilder.emptyTitleError',
     defaultMessage: 'Provide a question title',
   },
   errorMessage: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.errorMessage',
+    id: 'app.components.formBuilder.errorMessage',
     defaultMessage:
       'There is a problem, please fix the issue to be able to save your changes',
   },
   emptyOptionError: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.emptyOptionError',
+    id: 'app.components.formBuilder.emptyOptionError',
     defaultMessage: 'Provide at least 1 answer',
   },
   logicValidationError: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.logicValidationError',
+    id: 'app.components.formBuilder.logicValidationError',
     defaultMessage: 'Logic may not link to prior pages',
   },
   page: {
-    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.Page',
+    id: 'app.components.formBuilder.Page',
     defaultMessage: 'Page',
   },
 });
