+++ front/app/containers/Admin/projects/project/inputForm/utils.tsx
@@ -16,8 +16,4 @@ export const ideationConfig: FormBuilderConfig = {
   showStatusBadge: false,
   isLogicEnabled: false,
   isEditPermittedAfterSubmissions: true,
-
-  getGoBackUrl: (projectId: string) => {
-    return `/admin/projects/${projectId}/ideaform`;
-  },
 };
